Refresh the token if it has expired
The token we get is valid for 3600 seconds. Since we're not an app that
acts on behalf as the user when they're offline we don't get a refresh
token (and that would be pointless anyway since we only have session
storage).

This ensures that when a ClientIdError is raised we clear the session
and redirect, triggering the OAuth flow again and getting a new token.
Since the user has already given consent this is barely noticeable.