ui: Ensure people can log out
Due to the use of the `@login_required` decorator, which checks that you
come from a whitelisted domain, people that were logged in without
coming from a whitelisted domain would never be able to logout.

Instead, check if we have a valid Google session, revoke the token and
clear the session and retrigger the OAuth flow.

When we're not logged in it'll clear the session for good meassure and
retrigger the OAuth flow.